requirements.txt
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-texttospeech==2.26.0
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
idna==3.10
proto-plus==1.26.1
protobuf==6.31.0rc1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyPDF2==3.0.1
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
urllib3==2.4.0
